update 10.22
general bug fixes, lock visual changes, mission tracker system, Leo mech, and more

+++ Assets/Scripts/MissionSystem/DisableMTAlert.cs
@@ -7,7 +7,7 @@ public class DisableMTAlert : MonoBehaviour
     [SerializeField]
     string Marker;
     [SerializeField]
-    int Amount = 1;
+    float Amount = 1;
